finished CRUD API on screen / graph

+++ app/model/dashboard/dashboard_graph.go
@@ -21,10 +21,10 @@ type DashboardGraph struct {
 	Hosts      string `json:"hosts" gorm:"column:hosts"`
 	Counters   string `json:"counters" gorm:"column:counters"`
 	ScreenId   int64  `json:"screen_id" gorm:"column:screen_id"`
-	TimeSpan   int    `json:"timespan" gorm:"column:timespan"`
+	TimeSpan   int64  `json:"timespan" gorm:"column:timespan"`
 	GraphType  string `json:"graph_type" gorm:"column:graph_type"`
 	Method     string `json:"method" gorm:"column:method"`
-	Position   int    `json:"position" gorm:"column:position"`
+	Position   int64  `json:"position" gorm:"column:position"`
 	FalconTags string `json:"falcon_tags" gorm:"column:falcon_tags"`
 }
